Tugas Akhir/requirements.txt
argon2-cffi
PyJWT
sqlite3
